+++ runway/__init__.py
@@ -1,2 +1,2 @@
 """Set package version."""
-__version__ = '0.11.1'
+__version__ = '0.12.0rc1'